criptfs: Fix boot loop issue with encrypted userdata          due to stack corruption.
ioctl BLKGETSIZE expects pointer to unsigned long as argument.
If compiled with Nougat(7.1) sources for 64bit target using pointer to unsigned int causes stack
corruption due to type mismatch and TWRP boot loops on splash screen with this massage in recovery.log:
  "...
   get_crypt_ftr_info crypto key location: 'footer'
   stack corruption detected
   ..."

props to https://github.com/aopp/android_system_vold/commit/f8b8787317fc94439b63bc891eeda83f7ae2f4f6

+++ crypto/lollipop/cryptfs.c
@@ -713,13 +713,13 @@ static void get_device_scrypt_params(struct crypt_mnt_ftr *ftr) {
 
 static unsigned int get_blkdev_size(int fd)
 {
-  unsigned int nr_sec;
+  unsigned long nr_sec;
 
   if ( (ioctl(fd, BLKGETSIZE, &nr_sec)) == -1) {
     nr_sec = 0;
   }
 
-  return nr_sec;
+  return (unsigned int) nr_sec;
 }
 
 static int get_crypt_ftr_info(char **metadata_fname, off64_t *off)
